fix(roc-package-web-app-react): Pre-transpile universal-config
This will make it easier when testing amongst other things.
